Switched DataContainer Dimensions to 'size_t' which will vary based on how DREAM3D is compiled. The writing to HDF5 files is still with 64 bit integers. This part of the code compiles with a few hundred warnings. Added also some sanity checks for some of the data file readers to make sure we are not going to try and read more data that can fit into our address space.

+++ Source/DREAM3DLib/IO/DxWriter.cpp
@@ -84,7 +84,7 @@ int DxWriter::writeFile()
   GET_NAMED_ARRAY_SIZE_CHK_NOMSG_RET(m, Voxel, DREAM3D::VoxelData::GrainIds, Int32ArrayType, int32_t, totalPoints, grain_indicies);
 
   int err = 0;
-  int64_t dims[3];
+  size_t dims[3];
   m->getDimensions(dims);
   // std::cout << "Write Dx Grain File:  x, y, z: " << dims[0] << " " << dims[1] << " " << dims[2] << std::endl;
